requirements.txt
torch
torchvision
torchaudio

# Other dependencies
scipy
numpy
pandas
tqdm
pyyaml
matplotlib
pywavelets
ssqueezepy
tensorboard
mlflow